added protecting to goal routes in goalRoutes.js. Added a user check to prevent the wrong user from deleting or updating another users goal in goalControllers.js

+++ backend/controllers/goalControllers.js
@@ -2,11 +2,13 @@
 const asyncHandler = require('express-async-handler')
 //Instance for mongoose models to prefrom tasks
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 // @desc Get Goals
 // @route GET /api/goals
 // @access Private
 const getGoal = asyncHandler(async (req,res)=>{
-    const goals = await Goal.find({})
+
+    const goals = await Goal.find({user: req.user.id})
     //waiting to retrieve data in the form of an obj.
     res.status(200).json(goals)
 })
@@ -22,6 +24,8 @@ const setGoal = asyncHandler(async (req,res)=>{
     }
     const goal = await Goal.create({
         text: req.body.text,
+        //need to add user after authorization step
+        user: req.user.id,
     })
     res.status(200).json(goal)
 })
@@ -29,10 +33,23 @@ const setGoal = asyncHandler(async (req,res)=>{
 // @route PUT /api/goals/:id
 // @access Private
 const updateGoal = asyncHandler(async (req,res)=>{
+
     const goal = await Goal.findById(req.params.id)
     if(!goal){
         res.status(400)
         throw new Error ('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    //check for a user to associate with
+    if(!user){
+        res.status(401)
+        throw new Error('User Not Found')
+    }
+    //Make sure the logged in user(user.id) matches the goal user.
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('Not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -51,6 +68,18 @@ const deleteGoal = asyncHandler(async (req,res)=>{
         res.status(400)
         throw new Error('No Goal found with that ID')
     }
+    const user = await User.findById(req.user.id)
+    //check for a user to associate with
+    if(!user){
+        res.status(401)
+        throw new Error('User Not Found')
+    }
+    //Make sure the logged in user(user.id) matches the goal user.
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('Not authorized')
+    }
+    
     await goal.deleteOne()
     res.status(200).json({id:req.params.id})
 })
